pyproject.toml
[project]
name = "yt-to-blog"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "ollama>=0.4.8",
    "python-dotenv>=1.1.0",
    "youtube-transcript-api>=1.0.3",
]
